cleanup & fmt
