+++ client/src/lib.rs
@@ -20,6 +20,7 @@ cfg_if! {
     } else {}
 }
 
+mod conditioned_packet_receiver;
 mod error;
 mod impls;
 mod io;
