+++ server/src/lib.rs
@@ -16,6 +16,7 @@ extern crate log;
 #[macro_use]
 extern crate cfg_if;
 
+mod conditioned_packet_receiver;
 mod error;
 mod impls;
 mod io;
